Refactor expression implementation
